feat: 제너릭 처리

+++ app/src/main/kotlin/org/example/enums/Operator.kt
@@ -2,21 +2,21 @@ package org.example.enums
 
 enum class Operator(val symbol: String) {
     PLUS("+") {
-        override fun apply(num1: Int, num2: Int): Int = num1 + num2
+        override fun apply(num1: Double, num2: Double): Double = num1 + num2
     },
     MINUS("-") {
-        override fun apply(num1: Int, num2: Int): Int = num1 - num2
+        override fun apply(num1: Double, num2: Double): Double = num1 - num2
     },
     MULTIPLY("*") {
-        override fun apply(num1: Int, num2: Int): Int = num1 * num2
+        override fun apply(num1: Double, num2: Double): Double = num1 * num2
     },
     DIVIDE("/") {
-        override fun apply(num1: Int, num2: Int): Int? {
-            return if (num2 != 0) num1 / num2 else null
+        override fun apply(num1: Double, num2: Double): Double? {
+            return if (num2 != 0.0) num1 / num2 else null
         }
     };
 
-    abstract fun apply(num1: Int, num2: Int): Int?
+    abstract fun apply(num1: Double, num2: Double): Double?
 
     companion object {
         fun fromSymbol(symbol: String): Operator? {
